+++ services/gateway/__tests__/types/users.d.ts
@@ -1,0 +1,4 @@
+export interface ILoginSuccessResponse {
+  refreshToken: string;
+  eol: number;
+}