+++ services/users/__tests__/jest.config.default.ts
@@ -3,11 +3,14 @@ import { defaults } from 'jest-config';
 
 const config: Config = {
   verbose: true,
-  moduleDirectories: ['node_modules', 'src'],
-  moduleFileExtensions: [...defaults.moduleFileExtensions, 'mts', 'ts'],
+  moduleFileExtensions: [...defaults.moduleFileExtensions, 'ts'],
   testPathIgnorePatterns: ['build'],
   preset: 'ts-jest',
+  testMatch: ['**/*.test.ts'],
+  transformIgnorePatterns: ['<rootDir>/node_modules/(?!(mongodb-memory-server/index.d.ts))'],
   testEnvironment: 'node',
+  forceExit: true,
+  clearMocks: true,
 };
 
 export default config;