+++ services/users/jest.config.ts
@@ -1,9 +1,0 @@
-import { Config } from "jest";
-import defaultConfig from "./jest.config.default";
-
-const config: Config = {
-  ...defaultConfig,
-  roots: ["./tests"]
-};
-
-export default config;